chore: update pipeline and base from experimental
