+++ test/test_files/yaml/test_pipeline.yaml
@@ -2,11 +2,11 @@ components:
   Comp1:
     init_parameters:
       an_init_param: null
-    type: test.core.pipeline.test_pipeline.FakeComponent
+    type: test.core.pipeline.test_pipeline_base.FakeComponent
   Comp2:
     init_parameters:
       an_init_param: null
-    type: test.core.pipeline.test_pipeline.FakeComponent
+    type: test.core.pipeline.test_pipeline_base.FakeComponent
 connections:
 - receiver: Comp2.input_
   sender: Comp1.value
